chore: update yargs version to avoid noise in compliance system (#2512)
#### Description of changes

Some of our dependencies use outdated versions of yargs that are vulnerable to [CVE-2020-7608](https://cve.mitre.org/cgi-bin/cvename.cgi?name=CVE-2020-7608). This CVE does not impact us significantly because we only have transitive dependencies on yargs via devDependencies, but the CVE causes noise in our compliance systems, so updating to reduce the noise.

The affected root dependencies are webpack-cli, license-check-and-add, and node-sass. I verified that `yarn fastpass`, `yarn build`, and `npx webpack --help` work the same with this PR as without.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS