feat(notice): generate and embed notice file with popup hamburger menu link (#2525)
#### Description of changes

This PR adds a new menu item to our popup page's hamburger menu to show a "Third party notices" page that is embedded in the product. It sets up our CI build to automatically generate an up to date notice file with each build and embed that in canary/playground/insider/prod builds, and uses a placeholder file for local dev builds in the same format as the generated files.

![screenshot highlighting new menu option](https://user-images.githubusercontent.com/376284/79896209-eeb2b180-83bc-11ea-8f3c-3591714a9d98.png)

![screenshot of embedded notice page](https://user-images.githubusercontent.com/376284/79896275-05f19f00-83bd-11ea-83dc-148aa1a40a77.png)

A build with a "full" notice file has been published to [playground](https://chrome.google.com/webstore/detail/accessibility-insights-fo/oiokpkjanjfndgbdgepalohbolalekek) already (v2020.4.21.48), based on 99c1b7b, feel free to play with it live there.

Opening as a draft for now because I'm not sure how the PR build in the ms devops account will respond to the new task, which requires an extension from the MS open source programs office for generating the notice file that I've installed to our internal release deveops account but not the the public one; will probably have to iterate a bit for the PR build to pass.

I ran the HTML-formatted notice file in playground through a full AI4Web assessment already; the only issue I noticed was the `<h2>` near the top that ought to be a `<strong>` instead. Already filed a bug on the generator for that; out of scope for this PR.

We think a fuller solution would be to make a new "about" panel to contain this link that Unified could also use, but defering that for a separate PR because wiring up a new side panel is substantially more of a change than adding a new menu item here.

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS