fix: Move Info & Examples title from content to panel header (#2572)
Issue #2380 (Info & Examples panel close button is misaligned) calls out that the close button in the info & examples panel is positioned inconsistently with our other panels. This happens because the titles for the info & examples are included in the content of the panel rather than as the header. To address the issue, this PR moves all info & examples titles to the header of the panel. ContentCreatorWithTitle and ContentPageConfiguration are introduced to accomplish this. Credit to @dbjorge for initial investigation work and the concept of the change!

Note that this PR changes a lot of infoAndExamples files. The GitHub diff shows these files as heavily modified. This is because separating the title out changes the indentation level of the entire file. The only intended/noteworthy changes to those files occur in the first 10 or so lines, where the titles are pulled out of the content.