chore: Add test to verify infoAndExamples pageTitles match requirement names (#2595)
PR #2380 introduced pageTitle to infoAndExamples content and updated all existing content to have a pageTitle via createWithTitle. Unfortunately, as seen in PR #2593, it is easy to not set the pageTitle for infoAndExamples by using create. This PR introduces a test in the style of the assessments-style-cop test which verifies that all infoAndExamples content (of all requirements of all assessments) has a pageTitle value that matches its requirement's name. This test will fail if infoAndExamples content is created using create rather than createWithTitle.

A requirement name and a few infoAndExamples pageTitles were updated to pass this test.