fix: auditory cues to convey information are WCAG 1.1.1 not 1.3.3
WCAG 2.1 1.3.3. is only scope to *instructions* that describe other things based on sensory characteristics. The use of audio alone as a cue is not a failure of 1.3.3, but of 1.1.1. Audio-only cues are
non-text content (and time-based media, though relatively short ones). This may not be immediately clear/obvious from the normative wording of the SC (nor the current normative definition of "non-text content"
which gives only visual examples), but referring to 1.1.1 Understanding https://www.w3.org/WAI/WCAG21/Understanding/non-text-content.html see example 8:

> An e-learning application
>
> An e-learning application uses sound effects to indicate whether or not the answers are correct. The chime sound indicates that the answer is correct and the beep sound indicates that the answer is incorrect.
> A text description is also included so that people who can't hear or understand the sound understand whether the answer is correct or incorrect.