fix: 1.4.10 Reflow is not solely about text
Normatively, 1.4.10 is scoped to all content, not just text. While the principal concern of the SC is indeed text, it applies in equal measure to other types of content/mixed content.
Admittedly, the normative wording is a bit wooly/omits certain extra scenarios (i.e. that the primary concern is that blocks of text are legible without tow-dimensional scrolling, rather than an outright
ban on two-dimensional scrolling ... see for instance https://github.com/w3c/wcag/issues/668), but one things it does not do is explicitly limit this to just the text portion of content.

> Success Criterion 1.4.10 Reflow (Level AA): *Content* can be presented without loss of information or functionality [...]
(emphasis on "Content", not "Text" or "Text content")

https://www.w3.org/WAI/WCAG21/Understanding/reflow.html