feat(android-bug-filing): Introduce OpenIssueLink to remove BrowserAdapter from createFileIssueHandler (#2726)
In preparation for issue filing in the unified client, we need to remove the issue filing dependence on the web-only BrowserAdapter. This PR accomplishes that by creating OpenIssueLink and using it in place of BrowserAdapter.createActiveTab. There should be no impact on functionality. A future PR will use this to enable issue filing on the unified side.