feat(android-bug-filing): Get ToolData at moment issue is filed rather than initialization (#2733)
Currently, issue filing ToolData is set during initialization. This works for the web product, but the unified product's ToolData is not fully generated until a scan has been run (because it needs info from the android device). This PR refactors issue filing to get ToolData passed in along with other issue data at the moment an issue is filed. It should have no impact on functionality. A future PR will put it to use in the unified product.