fix(firefox-compat): avoid relative paths from background windows.create() (#2746)
#### Description of changes

Firefox and Chromium have different behavior when the webextension `windows.create()` API is invoked with a relative URL; Chromium resolves the relative URL relative to the root of the webextension, where Firefox resolves it relative to the path containing the invoking page. In practice, this means that Firefox tries to open DetailsView/debug pages as `moz-extension://extensionid/background/debug-tools/debug-tools.html` (note the extra `/background/` in there).

This resolves the firefox issue without impacting the chromium behavior by updating the cases that use relative URLs to instead use document-root-relative URLs, which both engines resolve as relative to the extension origin.

The combination of this, #2744, and #2745, and using Firefox 77 (beta) is enough to get a details view to load and mostly render in Firefox. There are still further compatibility issues out of scope for this PR (the menus in the popup view don't work, there are a few minor rendering issues in the requirements list, and all interactions with the target page seem broken)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: progress on #445
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS