feat(android-setup): Add core state machine logic (#2791)
* feat(android-setup): Add state machine

Given a set of steps (states), the job of the state machine is to link a step (state) to a set of actions. When an action is invoked on the state machine, the action is invoked for the current step. The step itself may fire a callback to transition to another step and the process repeats.

* feat(android-setup): Fix linting errors