chore: add CodeQL security scanning (#2803)
We discussed this with our engineering team in our weekly "engineering discussion" yesterday to make sure everyone would understand what was going on with the new checks. Folks were excited to give it a try, and unanimously agreed that it would be a good idea to enable it. Thanks again, Justin!