fix(android-setup): fix padding in prompt-connected-start-testing (#2947)
#### Description of changes

This PR adjusts the padding in the prompt-connected-start-testing step to match the figma.

It involves updates to a bunch of other components because previously, everything used a 24px margin between step layout header and content, but this particular fix requires a 16px margin. Oops.

I considered pushing back and just accepting a 24px margin, but it wasn't *that* big a deal to have it match the figma, and it does look a little nicer this way imo.

Verified no visual change on any other steps. Before and after of the impacted prompt-connected-start-testing step:

**Before**
![image](https://user-images.githubusercontent.com/376284/85476011-0d275b80-b56c-11ea-955f-145da2c52a80.png)

**After**
![image](https://user-images.githubusercontent.com/376284/85475908-cd607400-b56b-11ea-9eac-a5187ccf7bfb.png)


#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS