fix(android-setup): avoid detect-devices spinning forever on getDevices failure (#2959)
#### Description of changes

`getDevices` can fail, eg, if the device falls asleep or is disconnected during the call. If that happens, we don't want the step to spin forever; it should make *some* state transition. I think the ideal would be some sort of generic error step, but for now, the best one available is probably the one with the link about device connection troubleshooting, prompt-connect-to-device (same as the no-devices-connected case)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: from bug bash onenote
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS