fix(android-setup): fix aria-label on adb path text field (#2967)
#### Description of changes

Fluent UI has its own `ariaLabel` prop for `TextField`s, but not its own `aria-describedby`, so the casing has to be different for applying them (oops). React does provide an `aria-label` (which is why this wasn't a compile error), but Fluent ignores it in favor of its own property.

NVDA now reads the control (when empty) as "Enter a path edit Select the folder containing adb. We'll use it to connect to your device. adb path blank"

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS