fix(android-setup): fix scroll behavior with many devices/small size/high text zoom (#2974)
#### Description of changes

This PR implements the scrollbar behavior from the figma (also addresses a related bug bash onenote issue). Specifically, it:

* prevents the "choose device" list from showing its own scrollbar, in favor of a whole-dialog one
* fixes bottom buttons getting cut off when scrollbar is visible
* fixes outer scrollbar not being interactable (because it was rendering for a `-webkit-app-region: drag` component)

This PR is impacted by an issue where, in general, if a component inside a scrollable container is set to `-webkit-app-region: drag`, the scrollable container won't respond to mousewheel/middle-mouse scrolling when the mouse is over that inner component (instead, the window itself intercepts those events for drag handling). That means we had to choose whether to make the static background of the dialog window-draggable vs making it interact more normally with the scrollbar. Per offline discussion with @ferBonnin , we're going with the former for now, so this PR introduces a new layer of div (so the outer one with the scrollbar can be no-drag, enabling interaction with the scrollbar, but the inner one can still be drag).

This makes a minor update to the old device-connect-body for compatibility with the new container. It does not visually modify the old view; in particular, it neither fixes nor further regresses the pre-existing issues where the old view doesn't reflow appropriately to high text zoom / low window height.

Low-window-size choose device view, before (note multiple scrollbars and buttons cut off at bottom):
![image](https://user-images.githubusercontent.com/376284/85807988-6f1cc800-b708-11ea-8220-1e29b65c8067.png)

after (note single scrollbar only, with buttons at bottom not being cut off):
![image](https://user-images.githubusercontent.com/376284/85807968-61674280-b708-11ea-983d-ce9a17d0caae.png)

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [n/a] Addresses an existing issue: #0000
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS