fix: Firefox compatibility for script/CSS injection into target page (#3172)
#### Description of changes

This PR fixes the largest remaining blocker for Firefox support; it enables ai-web to successfully inject code into the target page. With this PR, the extension is mostly-functional in Firefox, though there are still several known issues outstanding that it does not address (see #445).

![screenshot of injected automated checks dialog rendering on a target page in Firefox](https://user-images.githubusercontent.com/376284/88836482-c220e980-d1a4-11ea-8319-5eefef786fd6.png)

The root cause of the incompatibility was this note buried in the documentation for the file parameter of the [`browser.tabs.insertCSS`](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/tabs/insertCSS) and [`browser.tabs.executeScript`](https://developer.mozilla.org/en-US/docs/Mozilla/Add-ons/WebExtensions/API/tabs/executeScript) APIs:

> **`file`** | *Optional*
>    `string`. Path to a file containing the code to inject. In Firefox, relative URLs are resolved relative to the current page URL. In Chrome, these URLs are resolved relative to the extension's base URL. To work cross-browser, you can specify the path as an absolute URL, starting at the extension's root, like this: `"/path/to/stylesheet.css"`.

This issue took a while to track down because the error message you get from Firefox when you ask it to inject a CSS/JS file with an invalid file path contains no useful information; it just says `Error: an unexpected error occurred`. We happen to use it here in a context where we're using `.then`/`.catch` callback handling all the way through our call stack, rather than `async`/`await`, which further exacerbated debugging due to not having full stack information from the error.

Technically, the only updates strictly required here were adding the `/` prefixes to the `cssFiles` and `jsFiles` in `content-script-injector.ts` and to the link elements in `shadow-initializer.ts`. The remaining changes are:

* Updates to `ContentScriptInjector` to log errors during injection to `logger.error` with context about which tab/file had trouble
* Updates to `WebextensionBrowserAdapter` to fail-fast with a descriptive error if a relative path would be passed to `insertCSS`/`executeScript`

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: Part of #445
- [x] Ran `yarn fastpass`
- [x] Added/updated relevant unit test(s) (and ran `yarn test`)
- [x] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [x] (UI changes only) Added screenshots/GIFs to description above
- [x] (UI changes only) Verified usability with NVDA/JAWS