feat(web-compliance): bundle LICENSE and NOTICE into unified packages (#3251)
#### Description of changes

This PR ensures that all of our platform distributions bundle all of the following LICENSE/NOTICE files:

* `LICENSE` (our own LICENSE, identical to the text that is bundled into each platform installer already)
* `NOTICE.html` (our NOTICE file, generated by Component Governance as of #3249)
* `LICENSE.electron.txt` (Electron's LICENSE, from the electron package; kept as-is per CELA guidance)
* `LICENSES.chromium.txt` (Chromium's LICENSES, from the electron package; kept as-is per CELA guidance)
* `resources/android-service/NOTICE.txt` ([Accessibility Insights for Android Service](https://github.com/microsoft/accessibility-insights-for-android-service)'s NOTICE file, to cover its transitive gradle dependencies)

#### Motivation

On all platforms, our electron packages are missing our NOTICE.html file. This bundles it into the application content root (on win/linux, the main install directory, on Mac, the "Content" directory under there)

Same for our LICENSE file (we do bundle it into the installers for each platform, but I also included it here as a file since it was the same process as NOTICE and we might want to include a link to it in a later "About" panel feature)

On Mac, long-outstanding issue https://github.com/electron-userland/electron-builder/issues/1495 means that we're *also* missing LICENSE.electron.txt and LICENSES.chromium.html. In the issue, the electron-builder maintainer [commented](https://github.com/electron-userland/electron-builder/issues/1495#issuecomment-296532959) that missing the chromium licenses "is ok" and closed the bug; our legal department disagrees and is very clear that we need to distribute the chromium licenses file alongside all our distributions.

#### Testing:

* [x] Linux: me to verify that the AppImage container contains all the files (it contains one additional/extra copy of our LICENSE file as "license.txt" because that's how electron-builder bundles it; we're okay with that)
* [x] Windows: me to verify that installing from the exe results in an installation directory containing all the files
* [x] Mac: @waabid to verify that installing from the dmg results in an installation directory whose Contents directory contains all the files

#### Pull request checklist
<!-- If a checklist item is not applicable to this change, write "n/a" in the checkbox -->
- [x] Addresses an existing issue: 1761407
- [x] Ran `yarn fastpass`
- [n/a] Added/updated relevant unit test(s) (and ran `yarn test`)
- [n/a] Verified code coverage for the changes made. Check coverage report at: `<rootDir>/test-results/unit/coverage`
- [x] PR title *AND* final merge commit title both start with a semantic tag (`fix:`, `chore:`, `feat(feature-name):`, `refactor:`). See `CONTRIBUTING.md`.
- [n/a] (UI changes only) Added screenshots/GIFs to description above
- [n/a] (UI changes only) Verified usability with NVDA/JAWS