feat(compliance-web): Address security warnings
Enabled security eslint rules for production code. Maintained disabling the below security eslint rules for code which is not released to production because of the extremely high false positive rate for the security rule set.

- (2) security/detect-non-literal-regexp
- (3) security/detect-non-literal-fs-filename
- (0) security/detect-unsafe-regex
- (0) security/detect-child-process
- (0) security/detect-eval-with-expression

Maintained the disabled rule 'security/detect-object-injection' because there are 295 warnings. Those will be addressed in another PR.